refactored queue_additive_rule
